Fix active divider detection
The code, in its current state, has a bug.  It determines whether or not to show/hide a cloned divider by comparing the absolute position of dividers with the *height* of a divider (`dividers[i].prop('offsetHeight')`).  This doesn't really make sense and explains the small-pixel level issues present in the plugin.

Instead, we want to compare the absolute position of the divider with the absolute position of the scrolling container (`$ionicPosition.offset($element).top`).